# Reduce allocations in aggregate