--- a/execution/model/operator.go
+++ b/execution/model/operator.go
@@ -11,10 +11,12 @@ import (
 
 // VectorOperator performs operations on series in step by step fashion.
 type VectorOperator interface {
-	// Next yields stream of samples representing series (vector of multiple series) per one or more steps.
+	// Next yields vectors of samples from all series for one or more execution steps.
 	Next(ctx context.Context) ([]StepVector, error)
 
-	// Series returns all series that we will see in all Next results.
+	// Series returns all series that the operator will process during Next results.
+	// The result can be used by upstream operators to allocate output tables and buffers
+	// before starting to process samples.
 	Series(ctx context.Context) ([]labels.Labels, error)
 
 	// GetPool returns pool of vectors that can be shared across operators.