--- a/execution/model/vector.go
+++ b/execution/model/vector.go
@@ -6,6 +6,8 @@ package model
 import "github.com/prometheus/prometheus/model/labels"
 
 type Series struct {
+	// ID is a numerical, zero-based identifier for a series.
+	// It allows using slices instead of maps for fast lookups.
 	ID     uint64
 	Metric labels.Labels
 }