This commit reduces allocations in the hash aggregate operator by recycling slices from the optional parameter operator and by reusing the same slice to store parameter arguments.

The commit also adds a benchmark for this case.

```
name                                                  old time/op    new time/op    delta
RangeQuery/vector_selector-8                            16.9ms ± 5%    17.4ms ± 6%     ~     (p=0.310 n=5+5)
RangeQuery/sum-8                                        12.0ms ±27%    10.9ms ± 3%     ~     (p=0.222 n=5+5)
RangeQuery/sum_by_pod-8                                 19.6ms ±39%    17.4ms ± 5%     ~     (p=0.310 n=5+5)
RangeQuery/rate-8                                       23.0ms ± 4%    23.0ms ± 1%     ~     (p=0.841 n=5+5)
RangeQuery/sum_rate-8                                   16.9ms ± 2%    17.3ms ± 4%     ~     (p=0.222 n=5+5)
RangeQuery/sum_by_rate-8                                23.9ms ±12%    23.0ms ± 1%     ~     (p=0.151 n=5+5)
RangeQuery/quantile_with_variable_parameter-8           40.1ms ±22%    28.4ms ± 3%  -29.23%  (p=0.008 n=5+5)
```